crypto.h : 1.	cambiado el nombre de la función para que sea con y griega 2.	iv y key cambiados de char a unsigned char
dataHolder
	cambiado size de long a int
crypto.c
	agregado. Por ahora solo funciona con DES

+++ wavEncrypt/include/dataHolder.h
@@ -5,10 +5,10 @@
 
 typedef struct {
 	void *data;
-	long size;
+	int size;
 } dataHolder_t;
 
-dataHolder_t newDataHolder(void *data, long size);
+dataHolder_t newDataHolder(void *data, int size);
 
 void dataHolderFree(dataHolder_t dataHolder);
